+++ mole/core/data.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 from dataclasses import dataclass, field
-from mole.common.help import FunctionHelper, InstructionHelper, SymbolHelper
+from mole.common.function import FunctionHelper
+from mole.common.instruction import InstructionHelper
+from mole.common.symbol import SymbolHelper
 from mole.common.log import log
 from mole.core.slice import (
     MediumLevelILBackwardSlicer,
