+++ mole/services/ai.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from concurrent import futures
 from datetime import datetime
-from mole.common.help import InstructionHelper
+from mole.common.instruction import InstructionHelper
 from mole.common.log import log
 from mole.common.task import BackgroundTask
 from mole.core.data import Path
